Refactor
added tests and src folders as decided by the gorup, works on #40
